mcp-server/pyproject.toml
[project]
name = "ebay-bluray-mcp"
version = "0.1.0"
description = "MCP server for managing eBay Blu-ray listings"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp>=1.0.0",
    "boto3>=1.34.0",
    "httpx>=0.25.0",
    "pillow>=10.1.0",
    "pandas>=2.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ebay_bluray_mcp"]

[project.scripts]
ebay-bluray-mcp = "ebay_bluray_mcp.server:main"
